refactor(e18e): split helpers.rs into helpers/ submodules
